requirements.txt
argparse
Pillow
